backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
supabase==2.9.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
requests==2.32.3
google-generativeai==0.8.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
Pillow==11.0.0
